feat: reove collectionList banner
